% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_core.R
\name{cache_location}
\alias{cache_location}
\title{Show the location of the msaVariant cache}
\usage{
cache_location()
}
\value{
The cache directory path (character).
}
\description{
Show the location of the msaVariant cache
}
